+++ banking-system/src/main/java/com/bankingsystem/models/Withdrawable.java
@@ -1,6 +1,8 @@
 package com.bankingsystem.models;
 
+import java.math.BigDecimal;
+
 public interface Withdrawable {
 
-    void withdraw(double amount);
+    void withdraw(BigDecimal amount);
 }
